+++ src/main/java/edu/ie3/simona/api/data/results/ExtResultDataConnection.java
@@ -10,17 +10,19 @@ import edu.ie3.datamodel.models.result.NodeResult;
 import edu.ie3.datamodel.models.result.ResultEntity;
 import edu.ie3.datamodel.models.result.system.SystemParticipantResult;
 import edu.ie3.simona.api.data.ExtOutputDataConnection;
+import edu.ie3.simona.api.data.ontology.DataMessageFromExt;
 import edu.ie3.simona.api.data.ontology.ScheduleDataServiceMessage;
 import edu.ie3.simona.api.data.results.ontology.ProvideResultEntities;
 import edu.ie3.simona.api.data.results.ontology.RequestResultEntities;
 import edu.ie3.simona.api.data.results.ontology.ResultDataMessageFromExt;
 import edu.ie3.simona.api.data.results.ontology.ResultDataResponseMessageToExt;
+import edu.ie3.simona.api.simulation.ontology.ControlResponseMessageFromExt;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.LinkedBlockingQueue;
-import org.apache.pekko.actor.ActorRef;
+import org.apache.pekko.actor.typed.ActorRef;
 
 /** Enables data connection of results between SIMONA and SimonaAPI */
 public class ExtResultDataConnection implements ExtOutputDataConnection {
@@ -30,13 +32,13 @@ public class ExtResultDataConnection implements ExtOutputDataConnection {
       new LinkedBlockingQueue<>();
 
   /** Actor reference to service that handles result data within SIMONA */
-  private ActorRef extResultDataService;
+  private ActorRef<DataMessageFromExt> extResultDataService;
 
   /** Actor reference to the dataServiceAdapter */
-  private ActorRef dataServiceActivation;
+  private ActorRef<DataMessageFromExt> dataServiceActivation;
 
   /** Actor reference to adapter that handles scheduler control flow in SIMONA */
-  private ActorRef extSimAdapter;
+  private ActorRef<ControlResponseMessageFromExt> extSimAdapter;
 
   /** Map uuid to external id of grid related entities */
   private final Map<UUID, String> gridResultAssetMapping;
@@ -59,7 +61,9 @@ public class ExtResultDataConnection implements ExtOutputDataConnection {
    * @param extSimAdapter actor ref to the extSimAdapter
    */
   public void setActorRefs(
-      ActorRef extResultDataService, ActorRef dataServiceActivation, ActorRef extSimAdapter) {
+      ActorRef<DataMessageFromExt> extResultDataService,
+      ActorRef<DataMessageFromExt> dataServiceActivation,
+      ActorRef<ControlResponseMessageFromExt> extSimAdapter) {
     this.extResultDataService = extResultDataService;
     this.dataServiceActivation = dataServiceActivation;
     this.extSimAdapter = extSimAdapter;
@@ -112,9 +116,9 @@ public class ExtResultDataConnection implements ExtOutputDataConnection {
    * @param msg the data/information that is sent to SIMONA's result data service
    */
   public void sendExtMsg(ResultDataMessageFromExt msg) {
-    extResultDataService.tell(msg, ActorRef.noSender());
+    extResultDataService.tell(msg);
     // we need to schedule data receiver activation with scheduler
-    extSimAdapter.tell(new ScheduleDataServiceMessage(dataServiceActivation), ActorRef.noSender());
+    extSimAdapter.tell(new ScheduleDataServiceMessage(dataServiceActivation));
   }
 
   /** Queues message from SIMONA that should be handled by the external simulation. */