+++ src/test/groovy/edu/ie3/simona/api/data/ev/ExtEvDataConnectionTest.groovy
@@ -1,38 +1,31 @@
 package edu.ie3.simona.api.data.ev
 
-import org.apache.pekko.actor.ActorSystem
-import org.apache.pekko.testkit.TestProbe
-import org.apache.pekko.testkit.javadsl.TestKit
 import edu.ie3.simona.api.data.ev.model.EvModel
-import edu.ie3.simona.api.data.ev.ontology.ProvideArrivingEvs
-import edu.ie3.simona.api.data.ev.ontology.ProvideCurrentPrices
-import edu.ie3.simona.api.data.ev.ontology.ProvideDepartingEvs
-import edu.ie3.simona.api.data.ev.ontology.ProvideEvcsFreeLots
-import edu.ie3.simona.api.data.ev.ontology.RequestCurrentPrices
-import edu.ie3.simona.api.data.ev.ontology.RequestDepartingEvs
-import edu.ie3.simona.api.data.ev.ontology.RequestEvcsFreeLots
+import edu.ie3.simona.api.data.ev.ontology.*
+import edu.ie3.simona.api.data.ontology.DataMessageFromExt
 import edu.ie3.simona.api.data.ontology.ScheduleDataServiceMessage
+import org.apache.pekko.actor.testkit.typed.javadsl.ActorTestKit
 import spock.lang.Shared
 import spock.lang.Specification
 
 class ExtEvDataConnectionTest extends Specification {
 
     @Shared
-    ActorSystem actorSystem
+    ActorTestKit testKit
 
     def setupSpec() {
-        actorSystem = ActorSystem.create()
+        testKit = ActorTestKit.create()
     }
 
     def cleanupSpec() {
-        TestKit.shutdownActorSystem(actorSystem)
-        actorSystem = null
+        testKit.shutdownTestKit()
+        testKit = null
     }
 
     def "ExtEvDataConnection should request and receive free evcs lots correctly"() {
         given:
-        def dataService = new TestProbe(actorSystem)
-        def extSimAdapter = new TestProbe(actorSystem)
+        def dataService = testKit.createTestProbe(DataMessageFromExt)
+        def extSimAdapter = testKit.createTestProbe(ScheduleDataServiceMessage)
         def extEvDataConnection = new ExtEvDataConnection()
         extEvDataConnection.setActorRefs(dataService.ref(), extSimAdapter.ref())
 
@@ -44,15 +37,15 @@ class ExtEvDataConnectionTest extends Specification {
         def actualReceivedEvcs = extEvDataConnection.requestAvailablePublicEvcs()
 
         then:
-        dataService.expectMsg(new RequestEvcsFreeLots())
-        extSimAdapter.expectMsg(new ScheduleDataServiceMessage(dataService.ref()))
+        dataService.expectMessage(new RequestEvcsFreeLots())
+        extSimAdapter.expectMessage(new ScheduleDataServiceMessage(dataService.ref()))
         actualReceivedEvcs == sentMsg.evcs()
     }
 
     def "ExtEvDataConnection should request and receive current charging prices correctly"() {
         given:
-        def dataService = new TestProbe(actorSystem)
-        def extSimAdapter = new TestProbe(actorSystem)
+        def dataService = testKit.createTestProbe(DataMessageFromExt)
+        def extSimAdapter = testKit.createTestProbe(ScheduleDataServiceMessage)
         def extEvDataConnection = new ExtEvDataConnection()
         extEvDataConnection.setActorRefs(dataService.ref(), extSimAdapter.ref())
 
@@ -64,15 +57,15 @@ class ExtEvDataConnectionTest extends Specification {
         def actualReceivedPrices = extEvDataConnection.requestCurrentPrices()
 
         then:
-        dataService.expectMsg(new RequestCurrentPrices())
-        extSimAdapter.expectMsg(new ScheduleDataServiceMessage(dataService.ref()))
+        dataService.expectMessage(new RequestCurrentPrices())
+        extSimAdapter.expectMessage(new ScheduleDataServiceMessage(dataService.ref()))
         actualReceivedPrices == sentMsg.prices()
     }
 
     def "ExtEvDataConnection should request and receive departing EVs correctly"() {
         given:
-        def dataService = new TestProbe(actorSystem)
-        def extSimAdapter = new TestProbe(actorSystem)
+        def dataService = testKit.createTestProbe(DataMessageFromExt)
+        def extSimAdapter = testKit.createTestProbe(ScheduleDataServiceMessage)
         def extEvDataConnection = new ExtEvDataConnection()
         extEvDataConnection.setActorRefs(dataService.ref(), extSimAdapter.ref())
 
@@ -86,15 +79,15 @@ class ExtEvDataConnectionTest extends Specification {
         def actualReceivedEvs = extEvDataConnection.requestDepartingEvs(requestedDepartingEvs)
 
         then:
-        dataService.expectMsg(new RequestDepartingEvs(requestedDepartingEvs))
-        extSimAdapter.expectMsg(new ScheduleDataServiceMessage(dataService.ref()))
+        dataService.expectMessage(new RequestDepartingEvs(requestedDepartingEvs))
+        extSimAdapter.expectMessage(new ScheduleDataServiceMessage(dataService.ref()))
         actualReceivedEvs == sentMsg.departedEvs()
     }
 
     def "ExtEvDataConnection should provide arriving EVs correctly"() {
         given:
-        def dataService = new TestProbe(actorSystem)
-        def extSimAdapter = new TestProbe(actorSystem)
+        def dataService = testKit.createTestProbe(DataMessageFromExt)
+        def extSimAdapter = testKit.createTestProbe(ScheduleDataServiceMessage)
         def extEvDataConnection = new ExtEvDataConnection()
         extEvDataConnection.setActorRefs(dataService.ref(), extSimAdapter.ref())
 
@@ -105,14 +98,14 @@ class ExtEvDataConnectionTest extends Specification {
         extEvDataConnection.provideArrivingEvs(arrivingEvs, Optional.of(60L))
 
         then:
-        dataService.expectMsg(new ProvideArrivingEvs(arrivingEvs, Optional.of(60L)))
-        extSimAdapter.expectMsg(new ScheduleDataServiceMessage(dataService.ref()))
+        dataService.expectMessage(new ProvideArrivingEvs(arrivingEvs, Optional.of(60L)))
+        extSimAdapter.expectMessage(new ScheduleDataServiceMessage(dataService.ref()))
     }
 
     def "ExtEvDataConnection should fail if wrong response is sent"() {
         given:
-        def dataService = new TestProbe(actorSystem)
-        def extSimAdapter = new TestProbe(actorSystem)
+        def dataService = testKit.createTestProbe(DataMessageFromExt)
+        def extSimAdapter = testKit.createTestProbe(ScheduleDataServiceMessage)
         def extEvDataConnection = new ExtEvDataConnection()
         extEvDataConnection.setActorRefs(dataService.ref(), extSimAdapter.ref())
 
@@ -124,8 +117,8 @@ class ExtEvDataConnectionTest extends Specification {
         extEvDataConnection.requestAvailablePublicEvcs()
 
         then:
-        dataService.expectMsg(new RequestEvcsFreeLots())
-        extSimAdapter.expectMsg(new ScheduleDataServiceMessage(dataService.ref()))
+        dataService.expectMessage(new RequestEvcsFreeLots())
+        extSimAdapter.expectMessage(new ScheduleDataServiceMessage(dataService.ref()))
         thrown RuntimeException
     }
 }