+++ src/test/groovy/edu/ie3/simona/api/data/primarydata/ExtPrimaryDataConnectionTest.groovy
@@ -1,19 +1,18 @@
 package edu.ie3.simona.api.data.primarydata
 
 import edu.ie3.datamodel.models.value.Value
+import edu.ie3.simona.api.data.ontology.DataMessageFromExt
 import edu.ie3.simona.api.data.ontology.ScheduleDataServiceMessage
 import edu.ie3.simona.api.data.primarydata.ontology.ProvidePrimaryData
 import edu.ie3.simona.api.test.common.DataServiceTestData
-import org.apache.pekko.actor.ActorSystem
-import org.apache.pekko.testkit.TestProbe
-import org.apache.pekko.testkit.javadsl.TestKit
+import org.apache.pekko.actor.testkit.typed.javadsl.ActorTestKit
 import spock.lang.Shared
 import spock.lang.Specification
 
 class ExtPrimaryDataConnectionTest extends Specification implements DataServiceTestData {
 
     @Shared
-    ActorSystem actorSystem
+    ActorTestKit testKit
 
     @Shared
     Map<String, UUID> extPrimaryDataMapping = Map.of(
@@ -22,18 +21,18 @@ class ExtPrimaryDataConnectionTest extends Specification implements DataServiceT
     )
 
     def setupSpec() {
-        actorSystem = ActorSystem.create()
+        testKit = ActorTestKit.create()
     }
 
     def cleanupSpec() {
-        TestKit.shutdownActorSystem(actorSystem)
-        actorSystem = null
+        testKit.shutdownTestKit()
+        testKit = null
     }
 
     def "ExtPrimaryDataConnection should provide primary data correctly"() {
         given:
-        def dataService = new TestProbe(actorSystem)
-        def extSimAdapter = new TestProbe(actorSystem)
+        def dataService = testKit.createTestProbe(DataMessageFromExt)
+        def extSimAdapter = testKit.createTestProbe(ScheduleDataServiceMessage)
         def extPrimaryDataConnection = new ExtPrimaryDataConnection(extPrimaryDataMapping)
         extPrimaryDataConnection.setActorRefs(
                 dataService.ref(),
@@ -50,14 +49,14 @@ class ExtPrimaryDataConnectionTest extends Specification implements DataServiceT
         extPrimaryDataConnection.providePrimaryData(0L, convertedPrimaryData, Optional.of(900L))
 
         then:
-        dataService.expectMsg(new ProvidePrimaryData(0L, convertedPrimaryData, Optional.of(900L)))
-        extSimAdapter.expectMsg(new ScheduleDataServiceMessage(dataService.ref()))
+        dataService.expectMessage(new ProvidePrimaryData(0L, convertedPrimaryData, Optional.of(900L)))
+        extSimAdapter.expectMessage(new ScheduleDataServiceMessage(dataService.ref()))
     }
 
     def "ExtPrimaryDataConnection should convert input data correctly"() {
         given:
-        def dataService = new TestProbe(actorSystem)
-        def extSimAdapter = new TestProbe(actorSystem)
+        def dataService = testKit.createTestProbe(DataMessageFromExt)
+        def extSimAdapter = testKit.createTestProbe(ScheduleDataServiceMessage)
         def extPrimaryDataConnection = new ExtPrimaryDataConnection(extPrimaryDataMapping)
         extPrimaryDataConnection.setActorRefs(
                 dataService.ref(),
@@ -69,14 +68,14 @@ class ExtPrimaryDataConnectionTest extends Specification implements DataServiceT
         extPrimaryDataConnection.convertAndSend(0L, inputDataMap, Optional.of(900L), log)
 
         then:
-        dataService.expectMsg(new ProvidePrimaryData(0L, Map.of(inputUuid, pValue), Optional.of(900L)))
-        extSimAdapter.expectMsg(new ScheduleDataServiceMessage(dataService.ref()))
+        dataService.expectMessage(new ProvidePrimaryData(0L, Map.of(inputUuid, pValue), Optional.of(900L)))
+        extSimAdapter.expectMessage(new ScheduleDataServiceMessage(dataService.ref()))
     }
 
     def "ExtPrimaryDataConnection should send no message, if input data for a not requested asset was provided"() {
         given:
-        def dataService = new TestProbe(actorSystem)
-        def extSimAdapter = new TestProbe(actorSystem)
+        def dataService = testKit.createTestProbe(DataMessageFromExt)
+        def extSimAdapter = testKit.createTestProbe(ScheduleDataServiceMessage)
         def extPrimaryDataConnection = new ExtPrimaryDataConnection(extPrimaryDataMapping)
         extPrimaryDataConnection.setActorRefs(
                 dataService.ref(),