+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased/Snapshot]
 
+### Changed
+- Converting pekko classic to typed [#232](https://github.com/ie3-institute/simonaAPI/issues/232)
+
 ## [0.7.0] - 2025-03-11
 
 ### Added